tile/elf: reorganize notify_exec()
This patch series addresses exe_file use for users that require the
mmap_sem for other things (mostly looking up the related vma->vm_file). 
In a lot of cases we end up with scenarios where we take the mmap_sem in
get_mm_exe_file(), just to then take it again soon after.  This is only
temporary as we will remove the need for mmap_sem when dealing with
exe_file.

This patch (of 3):

In the future mm->exe_file will be done without mmap_sem serialization,
thus isolate and reorganize the tile elf code to make the transition
easier.  Good users will, make use of the more standard get_mm_exe_file(),
requiring only holding the mmap_sem to read the value, and relying on
reference counting to make sure that the exe file won't dissappear
underneath us.

The visible effects of this patch are:

   o We now take and drop the mmap_sem more often. Instead of
     just in arch_setup_additional_pages(), we also do it in:

     1) get_mm_exe_file()
     2) to get the mm->vm_file and notify the simulator.

    [Note that 1) will disappear once we change the locking
     rules for exe_file.]

   o We avoid getting a free page and doing d_path() while
     holding the mmap_sem. This requires reordering the checks.

Signed-off-by: Davidlohr Bueso <dbueso@suse.de>
Cc: Chris Metcalf <cmetcalf@ezchip.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>