drivers/rtc/rtc-mc13xxx.c: fix obfuscated and wrong format string
According to C99, %2.s means 'print two spaces' (a precision of .  without
following digits or * means 0).  The kernel's printf implementation,
however, treats that case as if no precision was given, but relying on
that quirk is rather silly.  Also, since no - (aka left-justify) flag is
given, the field with of 2 would then cause the alarm->enabled case to
come out as "o n".  Deobfuscate it.

Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Cc: Alessandro Zummo <a.zummo@towertech.it>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>