hugetlbfs: accept subpool min_size mount option and setup accordingly
Make 'min_size=' be an option when mounting a hugetlbfs.  This option
takes the same value as the 'size' option.  min_size can be specified with
specifying size.  If both are specified, min_size must be less that or
equal to size else the mount will fail.  If min_size is specified, then at
mount time an attempt is made to reserve min_size pages.  If the
reservation fails, the mount fails.  At umount time, the reserved pages
are released.

Signed-off-by: Mike Kravetz <mike.kravetz@oracle.com>
Cc: Davidlohr Bueso <dave@stgolabs.net>
Cc: Aneesh Kumar <aneesh.kumar@linux.vnet.ibm.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Andi Kleen <andi@firstfloor.org>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>