mm: meminit: make __early_pfn_to_nid SMP-safe and introduce meminit_pfn_in_nid
__early_pfn_to_nid() in the generic and arch-specific implementations use
static variables to cache recent lookups.  Without the cache boot times
are much higher due to the excessive memblock lookups but it assumes that
memory initialisation is single-threaded.  Parallel initialisation of
struct pages will break that assumption so this patch makes
__early_pfn_to_nid() SMP-safe by requiring the caller to cache recent
search information.  early_pfn_to_nid() keeps the same interface but is
only safe to use early in boot due to the use of a global static variable.
 meminit_pfn_in_nid() is an SMP-safe version that callers must maintain
their own state for.

Signed-off-by: Mel Gorman <mgorman@suse.de>
Cc: Daniel J Blueman <daniel@numascale.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Nathan Zimmer <nzimmer@sgi.com>
Cc: Robin Holt <holt@sgi.com>
Cc: Scott Norton <scott.norton@hp.com>
Cc: Waiman Long <waiman.long@hp.com>
Cc: "Luck, Tony" <tony.luck@intel.com>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>